pyproject.toml

[project]
name = "dc_image_builder"
version = "0.0.0"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "aws-cdk-lib==2.206.0",
    "aws-cdk-aws-lambda-python-alpha==2.206.0a0",
    "certifi==2024.8.30",
    "constructs==10.4.2",
    "pyyaml==6.0.2",
]

[dependency-groups]
dev = [
    "ipdb==0.13.9",
    "ruff==0.6.7",
]

[tool.uv]
required-version = "<0.6.0"
package = false

[tool.uv.workspace]
members = ["lambda/update_base_images"]

[tool.ruff]
line-length = 80

[ruff.lint]
ignore = [
    "E501",
    "E203", #https://black.readthedocs.io/en/stable/the_black_code_style/current_style.html#slices
]
extend-select = [
    "I",
    "C4",
    "SIM",
    "Q003",
    "RET",
]
